Removed bgfx, renamed some things. fix: Somehow had the Mouse::MiddleButton and Mouse::RightButton values switched, now that's fixed

+++ VolcaniCore/src/VolcaniCore/Core/Entrypoint.cpp
@@ -7,7 +7,7 @@ extern Application* CreateApplication(const CommandLineArgs& args);
 
 int main(int argc, char** argv) {
 	Application::Init();
-	Application* app = CreateApplication(CommandLineArgs(argc, argv));
+	Application* app = CreateApplication(CommandLineArgs((uint32_t)argc, argv));
 	Application::Run();
 	Application::Close();
 }
